requirements-core.txt
# EchoLink Core Dependencies (without audio recording for initial development)

# Voice synthesis
elevenlabs>=1.0.0

# Text processing and summarization
openai>=1.0.0

# CLI and UI
rich>=13.0.0
keyboard>=0.13.5

# System monitoring and clipboard access
psutil>=5.9.0
pyperclip>=1.8.2

# File system monitoring
watchdog>=3.0.0

# HTTP requests for APIs
requests>=2.31.0

# Configuration management
python-dotenv>=1.0.0

# Audio processing (for playback)
pydub>=0.25.1

# JSON handling
jsonschema>=4.17.0 